# server/restAPI/docs/swagger.yaml
definitions:
  types.BackendData:
    properties:
      backend_name:
        type: string
      id:
        type: string
      plugin:
        type: string
      pointer:
        type: integer
    type: object
  types.JobData:
    properties:
      finish_time:
        type: string
      id:
        type: string
      metadata:
        additionalProperties: {}
        type: object
      pointer:
        type: integer
      qasm:
        type: string
      result_types:
        $ref: '#/definitions/types.JobResultTypes'
      results:
        $ref: '#/definitions/types.JobResultData'
      start_time:
        type: string
      status:
        type: string
      submission_date:
        type: string
      target_simulator:
        type: string
    type: object
  types.JobResultData:
    properties:
      counts:
        additionalProperties:
          type: number
        type: object
      expval:
        items:
          type: number
        type: array
      id:
        type: string
      job_id:
        type: string
      quasi_dist:
        additionalProperties:
          type: number
        type: object
    type: object
  types.JobResultTypes:
    properties:
      counts:
        type: boolean
      expval:
        type: boolean
      id:
        type: string
      job_id:
        type: string
      quasi_dist:
        type: boolean
    type: object
info:
  contact: {}
paths:
  /backend/{name}:
    get:
      description: backend data by backend name
      parameters:
      - description: Backend name
        in: path
        name: name
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/types.BackendData'
        "400":
          description: Invalid ID parameter
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: It wasn't possible to find the backend
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Failed during DB connection
          schema:
            additionalProperties:
              type: string
            type: object
      summary: get backend data
      tags:
      - backends
  /job/{id}:
    delete:
      description: delete all data related to this job id
      parameters:
      - description: Job ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Invalid ID parameter
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: No results for this ID
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Failed during DB connection
          schema:
            additionalProperties:
              type: string
            type: object
      summary: delete job data
      tags:
      - jobs
    get:
      description: get all data from job by ID
      parameters:
      - description: job ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/types.JobData'
        "400":
          description: Invalid ID parameter
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: It wasn't possible to find the job
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Failed during DB connection
          schema:
            additionalProperties:
              type: string
            type: object
      summary: get job data
      tags:
      - jobs
  /job/result/{id}:
    get:
      description: get job results by ID
      parameters:
      - description: Job ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/types.JobResultData'
        "400":
          description: Invalid ID parameter
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: No results for this ID
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Failed during DB connection
          schema:
            additionalProperties:
              type: string
            type: object
      summary: get job results
      tags:
      - jobs
  /jobs:
    get:
      description: get all data from jobs
      parameters:
      - description: Last id(order) gotten from db
        in: query
        name: cursor
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/types.JobData'
            type: array
        "500":
          description: Failed during DB connection
          schema:
            additionalProperties:
              type: string
            type: object
      summary: get jobs data
      tags:
      - jobs
  /plugin/{name}:
    delete:
      description: delete all data related to this plugin name
      parameters:
      - description: Plugin Name
        in: path
        name: name
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Invalid Name parameter
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: No results for this Name
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Failed during DB connection
          schema:
            additionalProperties:
              type: string
            type: object
      summary: delete plugin
      tags:
      - plugins
    post:
      description: add plugin by name
      parameters:
      - description: Plugin name as shown in the github org
        in: path
        name: name
        required: true
        type: string
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Invalid Name parameter
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: No results for this name
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Couldn't connect to database or get the plugin info from github
          schema:
            additionalProperties:
              type: string
            type: object
      summary: add plugin
      tags:
      - plugins
swagger: "2.0"